Create separate MeetOurTeam and AboutIntroSection components
